test-utils: split jest helpers into separate file
